+++ typings/eslint-plugin-chai-expect-keywords.d.ts
@@ -1,4 +1,0 @@
-declare module 'eslint-plugin-chai-expect-keywords' {
-  var chaiExpectKeywords: import('eslint').ESLint.Plugin
-  export = chaiExpectKeywords;
-}